Time: 19 ms (5.01%) | Memory: 32.1 MB (41.29%) - LeetSync

+++ 2631-sort-the-students-by-their-kth-score/sort-the-students-by-their-kth-score.cpp
@@ -5,7 +5,7 @@ public:
         for(size_t i=0; i<n-1; i++){
             for(size_t j=0; j<n-i-1; j++){
                 if(s[j][k] < s[j+1][k]){
-                    for(int l=0; l<m; l++){
+                    for(size_t l=0; l<m; l++){
                         swap(s[j][l], s[j+1][l]);
                     }
                 }
